+++ previous-versions/sync-for-expenses-version-1/RELEASES.md
@@ -39,3 +39,13 @@ Based on:
 - [python v0.4.0] previous-versions/sync-for-expenses-version-1
 ### Releases
 - [PyPI v0.4.0] https://pypi.org/project/codat-sync-for-expenses-version-1/0.4.0 - previous-versions/sync-for-expenses-version-1
+
+## 2023-10-02 00:09:55
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses-v1.yaml
+- Speakeasy CLI 1.92.3 (2.143.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.4.1] previous-versions/sync-for-expenses-version-1
+### Releases
+- [PyPI v0.4.1] https://pypi.org/project/codat-sync-for-expenses-version-1/0.4.1 - previous-versions/sync-for-expenses-version-1